Update CFG.md
Was not too happy with the earlier CFG. I tried writing this one from scratch but made the mistake of having it be
right-recursive from the get go. This didn't work too well. Would have made more sense to try making it left recursive
to begin with, maybe, and then change it. Anyway this is pretty much the same as EAC, with the exception that we have
assignment. That is handled mainly by performing 2 lookaheads and not by the grammar itself